rename model.pgsql to model.entity, create model.dto
