+++ management-server/src/main/kotlin/ut/isep/management/repository/pgsql/AssignmentRepository.kt
@@ -2,7 +2,7 @@ package ut.isep.management.repository.pgsql
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
-import ut.isep.management.model.pgsql.Assignment
+import ut.isep.management.model.entity.Assignment
 import java.util.*
