+++ management-server/src/main/kotlin/ut/isep/management/repository/pgsql/SectionRepository.kt
@@ -2,9 +2,8 @@ package ut.isep.management.repository.pgsql
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
-import ut.isep.management.model.pgsql.Applicant
-import ut.isep.management.model.pgsql.Assignment
-import ut.isep.management.model.pgsql.Section
+import ut.isep.management.model.entity.Assignment
+import ut.isep.management.model.entity.Section
 import java.util.*
