+++ management-server/src/test/kotlin/ut/isep/management/PostgreSQLRepositoryTests.kt
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
 import org.springframework.test.context.junit.jupiter.SpringExtension
-import ut.isep.management.model.pgsql.Applicant
+import ut.isep.management.model.entity.Applicant
 import ut.isep.management.repository.pgsql.ApplicantRepository
 
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
